requirements/base.txt
boto3==1.14.13
coreapi==2.3.3
ddt==1.2.2
dj-database-url==0.5.0
Django==2.2
django-allauth==0.36.0
django-cors-headers==2.4.0
django-countries==5.3.3
django-environ==0.4.5
django-extensions==2.1.4
django-import-export==1.2.0
django-rest-auth==0.9.3
django-silk==4.0.0
django-storages==1.6.6
djangorestframework==3.9.4
draftjs-exporter==3.0.0
factory-boy==2.12.0
gunicorn==19.9.0
Jinja2==2.10.1
MarkupSafe==1.1.0
pdfrw==0.4
psycopg2-binary==2.8.3
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.1
requests==2.22.0
six==1.11.0
stripe==2.0.3
termcolor==1.1.0

